big break
the raycast do not work in previous build because the camerarig does not following the vr hearset. Looks like bug(?) because the same test works in yesterday's build.
Solve the problem by changing the mainCamera reference to the CenterCamera(something like this, forgot the actual name) rather than the whole camerarig.
At first, I thought the issue is cause by mobile device performance because each raycast require 300 times of calculation, but even when i change the volume to 10, it still not work. so I try another approach.